refactor a bit; clean up main script in prep for moving class
